build(arch): infrastructure for Windows/WHP support
Update Cargo.toml with conditional dependencies and vm-memory git rev. Introduce src/arch/src/x86_64/linux/ and windows/ folder structures. Move existing Linux/KVM code to linux/ and add platform re-exports with cfg(target_os) conditional compilation.
