+++ src/arch/Cargo.toml
@@ -14,17 +14,26 @@ tdx = [ "tee", "dep:tdx" ]
 
 [dependencies]
 libc = ">=0.2.39"
-vm-memory = { version = "0.17", features = ["backend-mmap"] }
+vm-memory = { version = "0.17", default-features = false, features = ["backend-mmap"] }
 vmm-sys-util = "0.14"
 
 arch_gen = { package = "krun-arch-gen", version = "=0.1.0-1.18.0", path = "../arch_gen" }
-smbios = { package = "krun-smbios", version = "=0.1.0-1.18.0", path = "../smbios" }
 utils = { package = "krun-utils", version = "=0.1.0-1.18.0", path = "../utils" }
+
+[target.'cfg(not(target_os = "windows"))'.dependencies]
+smbios = { package = "krun-smbios", version = "=0.1.0-1.18.0", path = "../smbios" }
 
 [target.'cfg(target_os = "linux")'.dependencies]
 kvm-bindings = { version = "0.12", features = ["fam-wrappers"] }
 kvm-ioctls = "0.22"
 tdx = { version = "0.1.0", optional = true }
+
+[target.'cfg(target_os = "windows")'.dependencies]
+whp = { package = "krun-whp", version = "=0.1.0-1.18.0", path = "../whp" }
+windows-sys = { version = "0.61.2", features = [
+    "Win32_Foundation",
+    "Win32_System_Hypervisor",
+] }
 
 [dev-dependencies]
 utils = { package = "krun-utils", version = "=0.1.0-1.18.0", path = "../utils" }
