+++ src/arch/src/x86_64/windows/mod.rs
@@ -1,0 +1,2 @@
+pub mod msr;
+pub mod regs;
