+++ src/arch/src/x86_64/windows/msr.rs
@@ -1,0 +1,2 @@
+// Copyright 2026 Red Hat, Inc.
+// SPDX-License-Identifier: Apache-2.0
